+++ DuDoong-Api/src/main/java/band/gosrock/api/slack/handler/event/EventStatusChangeEventHandler.java
@@ -1,9 +1,10 @@
-package band.gosrock.api.event.handler;
+package band.gosrock.api.slack.handler.event;
 
 
 import band.gosrock.domain.common.alarm.EventSlackAlarm;
-import band.gosrock.domain.common.events.event.EventCreationEvent;
+import band.gosrock.domain.common.events.event.EventStatusChangeEvent;
 import band.gosrock.domain.domains.event.adaptor.EventAdaptor;
+import band.gosrock.domain.domains.event.domain.Event;
 import band.gosrock.domain.domains.host.adaptor.HostAdaptor;
 import band.gosrock.domain.domains.host.domain.Host;
 import band.gosrock.infrastructure.config.slack.SlackMessageProvider;
@@ -17,18 +18,19 @@ import org.springframework.transaction.event.TransactionalEventListener;
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class EventCreationEventHandler {
+public class EventStatusChangeEventHandler {
     private final HostAdaptor hostAdaptor;
     private final EventAdaptor eventAdaptor;
     private final SlackMessageProvider slackMessageProvider;
 
     @Async
     @TransactionalEventListener(
-            classes = EventCreationEvent.class,
+            classes = EventStatusChangeEvent.class,
             phase = TransactionPhase.AFTER_COMMIT)
-    public void handle(EventCreationEvent eventCreationEvent) {
-        final Host host = hostAdaptor.findById(eventCreationEvent.getHostId());
-        final String message = EventSlackAlarm.creationOf(eventCreationEvent.getEventName());
+    public void handle(EventStatusChangeEvent eventStatusChangeEvent) {
+        final Host host = hostAdaptor.findById(eventStatusChangeEvent.getHostId());
+        final Event event = eventAdaptor.findById(eventStatusChangeEvent.getEventId());
+        final String message = EventSlackAlarm.changeStatusOf(event);
 
         slackMessageProvider.sendMessage(host.getSlackUrl(), message);
     }
